+++ yaq_traits/_check.py
@@ -34,6 +34,11 @@ def check_trait(avpr, trait):
 
 def check(avpr):
     out = dict()
+    # check state defaults
+    for k, state in avpr.get("state", {}).items():
+        if "default" not in state:
+            raise Exception(f"state item \"{k}\" has no default")
+    # check traits
     for trait in traits.keys():
         out[trait] = check_trait(avpr, trait)
     return out